LPS-134215 Adjust Validation React Component to support the Date Validation
